Move decision scope into todos context
